Making binning of categorical data faster
Function that reduces a character list or numpy array to only n values. It keeps the n-1 most frequent occurrences as is in the list and sets the nth value to a new value called "other".

This is a connivence function to reduce the number of bins for categorical data.